# Speed up {Array,Hash}#hash and a couple of Hash operations