--- a/opal/corelib/array.rb
+++ b/opal/corelib/array.rb
@@ -1296,10 +1296,12 @@ class ::Array < `Array`
       }
 
       try {
-        values = $hash_ids.values();
-        for (item of values) {
-          if (#{eql?(`item`)}) {
-            return $opal32_add(result, 0x01010101);
+        if (!top) {
+          values = $hash_ids.values();
+          for (item of values) {
+            if (#{eql?(`item`)}) {
+              return $opal32_add(result, 0x01010101);
+            }
           }
         }