I have a use case where Hash#key? computing hashes for a not too large Hash/Array combo takes ~ 3 seconds, for the combination if things going on 10 seconds are wasted.
Using a Map instead of Object.create(null) for $hash_ids seems to reduce that to a 11ms for Hash#key? and the combination ~100ms or so.

Have not tried other use cases or anything else, let see result of the actions ...