@janbiedermann 2024-01-30
The huge speedup is due to error on my side, bit a more realistic ~10% speedup remains with things fixed, for my use case, using large Hash/Array combo as key with Hash#key?