@janbiedermann 2024-01-30
Another 10% speedup by avoiding function calls