@janbiedermann 2024-01-30
Another 10%, from 10s down to 7s.
Using the native Map iterator may benefit other places in Hash too