@janbiedermann 2024-01-30
Now down to below 7 seconds from 10 for the complete use case and down to 1.5s from 3s for Hash#key?, nice.